requirements.txt

langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
streamlit==1.45.1
PyPDF2==3.0.1
htbuilder
python-docx